Commiting before my laptop crashes again
