+++ NativeApp/src/components/Home/index.js
@@ -1,4 +1,4 @@
-import container from './Containers/container';
+import container from './container';
 import HomeView from './View/';
 
 export default container(HomeView);
