+++ NativeApp/src/screens/Login.js
@@ -1,7 +1,11 @@
 import React, { Component } from 'react';
+import {PropTypes as PT} from 'prop-types';
 import Login from '../components/Login';
 
 export default class LoginScreen extends Component {
+  static propTypes = {
+    navigation: PT.object,
+  }
 
   render() {
     return (
